SendGrid API refactoring - split classes
