+++ app/bundles/EmailBundle/Helper/PlainTextMassageHelper.php
@@ -1,0 +1,39 @@
+<?php
+
+/*
+ * @copyright   2015 Mautic Contributors. All rights reserved
+ * @author      Mautic
+ *
+ * @link        http://mautic.org
+ *
+ * @license     GNU/GPLv3 http://www.gnu.org/licenses/gpl-3.0.html
+ */
+
+namespace Mautic\EmailBundle\Helper;
+
+/**
+ * Class PlainTextMassageHelper.
+ */
+class PlainTextMassageHelper
+{
+    /**
+     * Extract plain text from message.
+     *
+     * @param \Swift_Message $message
+     *
+     * @return string
+     */
+    public static function getPlainTextFromMessage(\Swift_Message $message)
+    {
+        $children = (array) $message->getChildren();
+
+        foreach ($children as $child) {
+            $childType = $child->getContentType();
+            if ($childType === 'text/plain' && $child instanceof \Swift_MimePart) {
+                return $child->getBody();
+            }
+        }
+
+        return '';
+    }
+}
